+++ wpull/app_test.py
@@ -962,6 +962,22 @@ class TestApp(GoodAppTestCase):
 
         self.assertEqual(4, exit_code)
 
+    @wpull.testing.async.async_test(timeout=DEFAULT_TIMEOUT)
+    def test_database_uri(self):
+        arg_parser = AppArgumentParser()
+
+        with cd_tempdir():
+            args = arg_parser.parse_args([
+                self.get_url('/'),
+                '--database-uri', 'sqlite:///test.db'
+            ])
+
+            builder = Builder(args, unit_test=True)
+            app = builder.build()
+            exit_code = yield From(app.run())
+
+        self.assertEqual(0, exit_code)
+
 
 class SimpleHandler(tornado.web.RequestHandler):
     def get(self):