+++ wpull/options.py
@@ -226,11 +226,17 @@ class AppArgumentParser(argparse.ArgumentParser):
             metavar='FILE',
             help=_('load Lua hook script from FILE')
         )
-        group.add_argument(
+        database_group = group.add_mutually_exclusive_group()
+        database_group.add_argument(
             '--database',
             metavar='FILE',
             default=':memory:',
             help=_('save database tables into FILE instead of memory'),
+        )
+        database_group.add_argument(
+            '--database-uri',
+            metavar='URI',
+            help=_('save database tables at SQLAlchemy URI instead of memory'),
         )
         group.add_argument(
             '--concurrent',