+++ doc/changelog.rst
@@ -2,6 +2,9 @@
 What's New
 ==========
 
+* Added ``--database-uri`` option to specify a SQLAlchemy URI.
+
+
 0.1002 (2014-11-24)
 ===================